Use modulik to Hardware Wallets hot module replacement
